renesas-rcar: add a patch to improve wlan feature based on wl1837 module
Issue: LINUXEXEC-3167

There is a level translator(SN74LV1T04DBVR) between WL1837 and
gpio1-25. When there is an interrupt generated, WL1837 sends a
rasing edge from pin WLAN_IRQ, after through level translator, it is
a falling edge to GPIO1-25, so in dts node, setting interrupt type as
IRQ_TYPE_EDGE_FALLING is correct. But in kernel, there is code to
process the case that wifi module sends falling edge interrupt signal.
These code is not suitable for Kingfisher board, so get a patch from
SDK to fix this issue.

(LOCAL REV: NOT UPSTREAM) -- only for BSP renesas-rcar.

Signed-off-by: Meng Li <Meng.Li@windriver.com>