u-boot: Fix infrequent build failure, missing arch include link
On some build hosts were are seeing a u-boot build failure:

  boot/1_2017.09-r0/git/include/configs/ti_am335x_common.h:20:10: \
  fatal error: asm/arch/omap.h: No such file or directory

The omap.h header resides in an arch specific directory which is
linked into the 'build' directory. The makefile creates this link just
before doing the compilation of a file which requires the omap.h
header. Even in a failure case the log correctly indicates that the
link has been created, at the correct time, leaving one to believe the
issue is a deficiency in the build host's filesystem. To work around
this we instruct the Make to 'sync' the filesystem when the link can't
be found after it has been created.

Suggested-by: Mark Asselstine <mark.asselstine@windriver.com>

Issue: LIN1018-5107
Refer: http://lxgit.wrs.com/cgit/oe-core/commit/?h=WRLINUX_10_17_LTS&id=827dac9065f023c01f79a89652938c7bfca6ac41
Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>